ceph_test_objectstore: fix i386 build (again)
test/objectstore/store_test.cc: In member function ‘void SyntheticWorkloadState::read()’:
error: test/objectstore/store_test.cc:462:23: no matching function for call to ‘swap(uint64_t&, size_t&)’

+++ src/test/objectstore/store_test.cc
@@ -457,7 +457,7 @@ public:
     boost::uniform_int<> u1(0, max_object_len/2);
     boost::uniform_int<> u2(0, max_object_len);
     uint64_t offset = u1(*rng);
-    size_t len = u2(*rng);
+    uint64_t len = u2(*rng);
     if (offset > len)
       swap(offset, len);
